requirements.txt
mysql-connector-python==8.2.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
gemini==1.5.0
argparse
